% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_pool.R
\name{dash_aqerruy_generator}
\alias{dash_aqerruy_generator}
\title{Saldae dashboard header generator}
\usage{
dash_aqerruy_generator(dash_aqerruy = NULL)
}
\arguments{
\item{dash_aqerruy}{title ,author ,subtitle, Date(time),document type,theme}
}
\value{
YAML file containing header information
}
\description{
Saldae dashboard header generator
}
\author{
Farid Azouaou
}
