% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_pool.R
\name{dash_ul_generator}
\alias{dash_ul_generator}
\title{Saldae Reporting Engine: body part generator}
\usage{
dash_ul_generator(dash_ul = NULL, dash_params = NULL, report_output = NULL)
}
\arguments{
\item{dash_ul}{dashboard body information}

\item{dash_params}{dashboard parameters information}

\item{dash_layout}{dashboard layout object}
}
\value{
YAML file containing the body part of report
}
\description{
generates body part of rmarkdon object including layout and content
}
\author{
Farid Azouaou
}
