% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_pool.R
\name{.single_page_layout_generator}
\alias{.single_page_layout_generator}
\title{Saldae dashboard layout generator}
\usage{
.single_page_layout_generator(page_layout_configs = NULL)
}
\arguments{
\item{page_layout_configs}{layout config information of a single page /section}
}
\value{
page layout information
}
\description{
Saldae dashboard layout generator
}
\author{
Farid Azouaou
}
