% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_pool.R
\name{SALDAE_layout_generator}
\alias{SALDAE_layout_generator}
\title{Saldae dashboard layout generator}
\usage{
SALDAE_layout_generator(layout_configs = NULL)
}
\arguments{
\item{layout_configs}{t.b.d}
}
\value{
list of layout configurations
}
\description{
Saldae dashboard layout generator
}
\author{
Farid Azouaou
}
