% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saldae_publisher.R
\name{saldae_publihser_f}
\alias{saldae_publihser_f}
\title{Saldae Report Publisher}
\usage{
saldae_publihser_f(
  target_report = "Saldae_main_report.html",
  github_url = "git@github.com:fairanalytics/saldae_dashboard.git"
)
}
\arguments{
\item{github_url}{Github remote repository}

\item{target_repo}{target_repository}
}
\description{
Saldae Report Publisher
}
\author{
Farid Azouaou
}
