% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_pool.R
\name{SALDAE_reporting_engine}
\alias{SALDAE_reporting_engine}
\title{Saldae Reporting Engine}
\usage{
SALDAE_reporting_engine(
  dash_aqerruy = NULL,
  dash_asezwer = NULL,
  dash_ul = NULL,
  tisefka = NULL,
  output_file = NULL,
  dash_adhar = NULL,
  publish_report = FALSE
)
}
\arguments{
\item{dash_aqerruy}{dashboard header information}

\item{dash_ul}{dashboard body information}

\item{dash_adhar}{dashboard footer}
}
\value{
I DONT KNOW
}
\description{
a module which generate dynamic and customizable reports
}
