% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saldae_publisher.R
\name{update_publisher}
\alias{update_publisher}
\title{Generate or Update publisher(Web site)}
\usage{
update_publisher(new_report_name = NULL, repo_path = NULL)
}
\description{
Generate or Update publisher(Web site)
}
